requirements.txt
asyncpg
lxml
httpx
python-dotenv
redis[hiredis]
fastapi
uvicorn[standard]
sse-starlette
